requirements.txt
aiohttp==2.2.3
allpairspy==2.3.0
apipkg==1.4
Appium-Python-Client==0.24
async-timeout==1.2.1
asyncio==3.4.3
certifi==2017.7.27.1
chardet==3.0.4
execnet==1.4.1
idna==2.5
lxml==3.8.0
multidict==3.1.3
namedlist==1.7
py==1.4.34
pytest==3.1.3
pytest-allure-adaptor==1.7.7
pytest-xdist==1.18.2
requests==2.18.3
sauceclient==1.0.0
selenium==2.53.6
six==1.10.0
urllib3==1.22
yarl==0.12.0